splint into more components
