+++ src/lib/components/PuntoText.svelte
@@ -1,0 +1,7 @@
+<h1 class="punto fw-bold">
+	<span class="text-danger">p</span>
+	<span class="text-info">u</span>
+	<span class="text-warning">n</span>
+	<span class="text-success">t</span>
+	<span class="text-danger">o</span>
+</h1>
